Ensure that requirefires running in places with modified process globals inherit the process global
This is to match node's behaviour and specifically for jest, where requirefire is useful. Jest runs each suite in a node worker for test isolation, but before this change, the inner requirefired modules were escaping jest's process instance and using the root level process object. This ensures we get the same one as the context that creates requirefire.